backend/requirements.txt
asgiref==3.11.1
django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
mysql-connector-python==9.6.0
pyjwt==2.12.1
sqlparse==0.5.5
